+++ blakserv/blakres.h
@@ -15,7 +15,7 @@
 
 typedef struct resource_struct
 {
-   int resource_id;
+   __int64 resource_id;
    char *resource_val[MAX_LANGUAGE_ID];
    char *resource_name;
    struct resource_struct *next;
@@ -23,16 +23,16 @@ typedef struct resource_struct
 
 void InitResource(void);
 void ResetResource(void);
-void AddResource(int id, int lang_id, char *str_value);
-void SetResourceName(int id,char *name);
-int AddDynamicResource(char *str_value);
+void AddResource(__int64 id, int lang_id, char *str_value);
+void SetResourceName(__int64 id, char *name);
+__int64 AddDynamicResource(char *str_value);
 void ChangeDynamicResourceStr(resource_node *r,char *str_value);
 void ChangeDynamicResource(resource_node *r,char *data,int len_data);
 int GetNumDynamicRscFiles(void);
 int SetNumDynamicRscFiles(int num_files);
-resource_node * GetResourceByID(int id);
-char * GetResourceStrByLanguageID(int id, int lang_id);
-Bool IsResourceByID(int id);
+resource_node * GetResourceByID(__int64 id);
+char * GetResourceStrByLanguageID(__int64 id, int lang_id);
+Bool IsResourceByID(__int64 id);
 resource_node * GetResourceByName(const char *resource_name);
 void ForEachResource(void (*callback_func)(resource_node *r));
 void ForEachDynamicRsc(void (*callback_func)(resource_node *r));
